Move class out of minecraft package
